Work back with abstract class to split up between Organizer & Offer Controller.
