+++ app/Organizer/OrganizerControllerProvider.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Silex\Organizer;
 
 use CultuurNet\UDB3\Http\Import\RemoveEmptyArraysRequestBodyParser;
-use CultuurNet\UDB3\Http\Offer\GetPermissionsForCurrentUserRequestHandler;
-use CultuurNet\UDB3\Http\Offer\GetPermissionsForGivenUserRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\AddImageRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\AddLabelRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\DeleteAddressRequestHandler;
@@ -15,6 +13,8 @@ use CultuurNet\UDB3\Http\Organizer\DeleteImageRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\DeleteLabelRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\DeleteOrganizerRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\GetOrganizerRequestHandler;
+use CultuurNet\UDB3\Http\Organizer\GetPermissionsForCurrentUserRequestHandler;
+use CultuurNet\UDB3\Http\Organizer\GetPermissionsForGivenUserRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\ImportOrganizerRequestHandler;
 use CultuurNet\UDB3\Http\Organizer\LegacyOrganizerRequestBodyParser;
 use CultuurNet\UDB3\Http\Organizer\UpdateAddressRequestHandler;
@@ -67,8 +67,8 @@ class OrganizerControllerProvider implements ControllerProviderInterface, Servic
         $controllers->put('/{organizerId}/labels/{labelName}/', AddLabelRequestHandler::class);
         $controllers->delete('/{organizerId}/labels/{labelName}/', DeleteLabelRequestHandler::class);
 
-        $controllers->get('/{offerId}/permissions/', 'Organizer' . GetPermissionsForCurrentUserRequestHandler::class);
-        $controllers->get('/{offerId}/permissions/{userId}/', 'Organizer' . GetPermissionsForGivenUserRequestHandler::class);
+        $controllers->get('/{offerId}/permissions/', GetPermissionsForCurrentUserRequestHandler::class);
+        $controllers->get('/{offerId}/permissions/{userId}/', GetPermissionsForGivenUserRequestHandler::class);
 
         return $controllers;
     }
@@ -155,7 +155,7 @@ class OrganizerControllerProvider implements ControllerProviderInterface, Servic
             fn (Application $application) => new DeleteLabelRequestHandler($app['event_command_bus'])
         );
 
-        $app['Organizer' . GetPermissionsForCurrentUserRequestHandler::class] = $app->share(
+        $app[GetPermissionsForCurrentUserRequestHandler::class] = $app->share(
             function (Application $app) {
                 return new GetPermissionsForCurrentUserRequestHandler(
                     [Permission::organisatiesBewerken()],
@@ -165,7 +165,7 @@ class OrganizerControllerProvider implements ControllerProviderInterface, Servic
             }
         );
 
-        $app['Organizer' . GetPermissionsForGivenUserRequestHandler::class] = $app->share(
+        $app[GetPermissionsForGivenUserRequestHandler::class] = $app->share(
             function (Application $app) {
                 return new GetPermissionsForGivenUserRequestHandler(
                     [Permission::organisatiesBewerken()],
