+++ src/Http/Organizer/GetPermissionsForCurrentUserRequestHandler.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CultuurNet\UDB3\Http\Organizer;
+
+use CultuurNet\UDB3\Http\Permissions\GetPermissionsForCurrentUserRequestHandler as GetOrganizerPermissionsForCurrentUserRequestHandler;
+
+final class GetPermissionsForCurrentUserRequestHandler extends GetOrganizerPermissionsForCurrentUserRequestHandler
+{
+}
