+++ src/Http/Organizer/GetPermissionsForGivenUserRequestHandler.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CultuurNet\UDB3\Http\Organizer;
+
+use CultuurNet\UDB3\Http\Permissions\GetPermissionsForGivenUserRequestHandler as GetOrganizerPermissionsForGivenUserRequestHandler;
+
+final class GetPermissionsForGivenUserRequestHandler extends GetOrganizerPermissionsForGivenUserRequestHandler
+{
+}
